# Performance: Remove Extraneous Re-Renders in Frames Layer