## Context
Part of ongoing performance work.

## Summary
Removes empty re-renders of frame elements.

Running against these tests:
https://github.com/google/web-stories-wp/issues/10158

**Element Selection**
| Before | After |
| ----- | -----|
| React | React |
| Longest Commit | Longest Commit |
| ~92ms | ~56ms |
| Frames Layer | Frames Layer |
| ~45ms | ~1.7ms |
| <img width="630" alt="Screen Shot 2022-01-19 at 10 16 24 AM" src="https://user-images.githubusercontent.com/35983235/150170978-2882431f-f13c-4642-b82f-6c18a23ca415.png"> |  <img width="559" alt="Screen Shot 2022-01-19 at 10 16 09 AM" src="https://user-images.githubusercontent.com/35983235/150171006-5936ee78-929a-4196-8ace-39566a75232b.png"> |
| Total JS | Total JS |
| ~167ms | ~120ms |
| <img width="631" alt="Screen Shot 2022-01-19 at 10 28 02 AM" src="https://user-images.githubusercontent.com/35983235/150173019-b22d7ce1-6c74-46a9-b666-bb7e18300063.png"> |  <img width="500" alt="Screen Shot 2022-01-19 at 10 23 37 AM" src="https://user-images.githubusercontent.com/35983235/150173335-7acffd33-497d-4da8-a994-c9005ad2ddcb.png"> |






**Starting the drag of element to update position**
| Before | After |
| ----- | -----|
| React | React |
| Longest Commit | Longest Commit |
| ~80ms | ~2ms |
| Frames Layer | Frames Layer |
| ~44ms | ~1.6ms |
| <img width="686" alt="Screen Shot 2022-01-19 at 10 07 55 AM" src="https://user-images.githubusercontent.com/35983235/150170207-f43067f9-1695-4f3a-b130-152882468aa0.png"> |  <img width="679" alt="Screen Shot 2022-01-19 at 10 11 59 AM" src="https://user-images.githubusercontent.com/35983235/150170234-562b95b6-1af4-4ffd-85ab-79e3d6ee9222.png"> |
| Total JS | Total JS |
| ~182ms | ~36ms |
| <img width="687" alt="Screen Shot 2022-01-19 at 10 04 24 AM" src="https://user-images.githubusercontent.com/35983235/150169030-a1e895b2-323f-4dee-956f-2eb7ac43be81.png"> |  <img width="675" alt="Screen Shot 2022-01-19 at 10 05 04 AM" src="https://user-images.githubusercontent.com/35983235/150169055-410a50a5-904e-4a83-ac54-8bb641f86201.png"> |




**Ending the drag of element to update position**
| Before | After |
| ----- | -----|
| React | React |
| Longest Commit | Longest Commit |
| ~92ms | ~60ms |
| Frames Layer | Frames Layer |
| ~44ms | ~7.5ms |
| <img width="688" alt="Screen Shot 2022-01-19 at 9 54 59 AM" src="https://user-images.githubusercontent.com/35983235/150167370-8555e638-e44a-479b-86bc-0a9ba68f17c7.png"> | <img width="675" alt="Screen Shot 2022-01-19 at 9 51 20 AM" src="https://user-images.githubusercontent.com/35983235/150167401-89fc37dd-92ee-4989-8299-e404ea59801b.png"> |
| Total JS | Total JS |
| ~382ms | ~204ms |
| <img width="689" alt="Screen Shot 2022-01-19 at 10 01 13 AM" src="https://user-images.githubusercontent.com/35983235/150168485-dcfbbed6-c7c0-4110-bf68-5bcd0978f125.png"> |  <img width="678" alt="Screen Shot 2022-01-19 at 10 01 40 AM" src="https://user-images.githubusercontent.com/35983235/150168535-1d49b8ea-5b14-4082-ba9d-38c29eeab2ae.png"> |

**Changing Layer Order**
_Screenshots of flame graphs are similar to above, so going to omit for brevity here_
| Before | After |
| ----- | ----- |
| React | React |
| Longest Commit | Longest Commit |
| ~92ms | ~55ms |
| Frames Layer | Frames Layer |
| ~42ms | ~3.6ms |
| Total JS | Total JS|
| ~182ms | ~127ms |




## Relevant Technical Choices
Better context selector usage and memoization.

## To-do
NA
<!-- A list of things that need to be addressed in this PR or follow-up changes. -->

## User-facing changes
NA
<!--
Please describe your changes.
Include before/after screenshots or a short video.
-->

## Testing Instructions
NA
<!--
How can the changes in this PR be verified?
Please provide step-by-step instructions how to reproduce the issue, if applicable.
Write step-by-step test instructions aimed at non-tech-savvy users, even if the PR is not user-facing.
-->

<!-- ignore-task-list-start -->
- [x] This is a non-user-facing change and requires no QA
<!-- ignore-task-list-end -->

This PR can be tested by following these steps:

1.


## Reviews

### Does this PR have a security-related impact?

<!-- Examples: new APIs, changes to KSES, etc.  -->

### Does this PR change what data or activity we track or use?

<!-- Examples: changes to telemetry, new third-party APIs -->

### Does this PR have a legal-related impact?

<!-- Examples: new images with unknown sources, new production dependencies with incompatible licenses -->

## Checklist

<!-- Check these after PR creation -->

- [x] This PR addresses an existing issue and I have linked this PR to it in ZenHub
- [x] I have tested this code to the best of my abilities
- [x] I have verified accessibility to the best of my abilities ([docs](https://github.com/google/web-stories-wp/blob/main/docs/accessibility-testing.md))
- [x] I have verified i18n and l10n (translation, right-to-left layout) to the best of my abilities
- [x] This code is covered by automated tests (unit, integration, and/or e2e) to verify it works as intended ([docs](https://github.com/google/web-stories-wp/tree/main/docs#testing))
- [x] I have added documentation where necessary
- [x] I have added a matching `Type: XYZ` label to the PR

---

<!--
Please reference the issue(s) this PR addresses.
No URLs, just the issue numbers.
Use "Fixes #123" if it fixes an issue.

NOTE: One reference per line!

Example:

Fixes #123
Partially addresses #456
See #789
-->

Fixes #10079