@BrittanyIRL 2022-01-19
> > Testing this locally and found a lil bug from reducing the amount of rerenders:
> > 
> > * set media as your background
> > * use the quick action menu to replace that background media
> > * notice the layers panel icon has updated to the new media but the canvas has not
> 
> hmph I'm not able to replicate this consistently. Was able to get it to fail once, but was also able to get it to fail once on main as well. I don't think it's related to these issues.
> 
> Can you get it to fail consistently on your machine? if so, can you get the same issue to fail consistently on main?

So i can get it to fail on this branch and not on staging - it's not every time, try repeating the action a few times.