@littlemilkstudio 2022-01-20
@BrittanyIRL just to clarify, you're talking about this button right?
<img width="366" alt="Screen Shot 2022-01-20 at 8 09 12 AM" src="https://user-images.githubusercontent.com/35983235/150354412-ec581c2d-63db-43b6-8a77-b07abb2cfac3.png">