@littlemilkstudio 2022-01-20
okay interesting. able to get the image to fail loading in the icon and filters panel, but the canvas works fine for me