@BrittanyIRL 2022-01-20
> okay interesting. able to get the image to fail loading in the icon and filters panel, but the canvas works fine for me.
> But just checked and I'm also getting this when running main on my machine

Yes, so I meant that the icon on the layers panel updates but the canvas isn't - the bad image lingers. We chatted a bit off github so i'm just catching up here. You're totally right it's on main too - so false alarm. Let's keep tabs on this though - seems related to video. 

https://user-images.githubusercontent.com/10720454/150401895-afa0b092-4775-4010-8ab0-30e8b7703d65.mp4